+++ NEWS.md
@@ -7,11 +7,16 @@ shorthand will be renamed with a more explicit `"PurpleAir"`.
 
 - Renamed all `PurpleAir_~()` functions to `PurpleAir_~()`.
 - Renamed `PURPLE_AIR_API_~` keys to `PurpleAir_API_~`.
+- Renamed `purple_air_synoptic` class to `PurpleAir_synoptic`.
 - Added Clarity API functions:
   - `Clarity_getAllOpenHourly()`
   - `Clarity_getAllOpenIndividual()`
   - `Clarity_getOpenHourly()`
   - `Clarity_getOpenIndividual()`
+  - `Clarity_createSynoptic()`
+  - `Clarity_enhanceRawSynopticData()`
+- Added functions for non-PurpleAir synoptic data
+  - `syn_leaflet()`
 
 # AirSensor2 0.2.1